+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - When creating user login and email can be different now
 - Fixed Domain#add_user (GH issue #354)
 - Support for GoodData.connect ENV['GD_GEM_USER'], ENV['GD_GEM_PASSWORD']
+- Added Schedule#execute(:wait => true|false) option
 
 ## 0.6.11